Run PreFigure in a web worker
