+++ website/packages/playground/package.json
@@ -17,6 +17,7 @@
         "@types/file-saver": "^2.0.7",
         "bootstrap": "^5.3.3",
         "codemirror": "^6.0.1",
+        "comlink": "^4.4.2",
         "easy-peasy": "^6.0.5",
         "file-saver": "^2.0.5",
         "pyodide": "^0.26.4",
