+++ website/packages/playground/tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/App.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/editor.tsx","./src/components/renderer.tsx","./src/state/compiler.ts","./src/state/index.ts","./src/state/model.ts","./src/state/store.ts"],"errors":true,"version":"5.7.2"}
+{"root":["./src/App.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/editor.tsx","./src/components/renderer.tsx","./src/state/index.ts","./src/state/model.ts","./src/state/store.ts","./src/worker/compiler.ts","./src/worker/index.ts"],"version":"5.7.2"}
